SUN 1:04AM
Updated tentacle spritesheet, edited level to create a better path, moved animation loading into main. Changed how the spawner button works to allow for object to be added seperately. fixed path-following issue for the enemies, removed unneccessary properties such as move_speed.  Relocated the spawner timer code into the level base class, changed summon order for altar and menu background, edited spawner button as it was inaccessible behind the monster being spawned in (fixed the leaving behind of spawner button corpses). Created a collision detection system that compares two body ids to identify the objects they came from. Attempted to get the game to have a pause state, it didn't work, but there is a class still in the code from that.  Allowed the tentacles to attack and the enemies to take damage/die. created a game over state